Provider renaming: update basic code
Update provider name in basic code files.
