+++ selectel/regions_test.go
@@ -1,4 +1,4 @@
-package selvpc
+package selectel
 
 import (
 	"testing"
@@ -9,8 +9,8 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestExpandResellV2Regions(t *testing.T) {
-	r := resourceResellKeypairV2()
+func TestExpandVPCV2Regions(t *testing.T) {
+	r := resourceVPCKeypairV2()
 	d := r.TestResourceData()
 	d.SetId("1")
 	regions := []interface{}{"ru-1", "ru-2", "ru-3"}
@@ -18,7 +18,7 @@ func TestExpandResellV2Regions(t *testing.T) {
 
 	expected := []string{"ru-1", "ru-2", "ru-3"}
 
-	actual := expandResellV2Regions(d.Get("regions").(*schema.Set))
+	actual := expandVPCV2Regions(d.Get("regions").(*schema.Set))
 
 	assert.ElementsMatch(t, expected, actual)
 }
